Update chart widget for new patient analytics
